Refactoring of the BaseSelect. Select to HorizontalSelect
