+++ src/Select/BaseSelect.js
@@ -1,31 +1,50 @@
-import React from 'react'
+import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import styles from './BaseSelect.module.scss'
 
-const BaseSelect = ({ selectedIndex, onClick, options, ...props }) => {
-  return (
-    <div className={styles.wrapper} {...props}>
-      {options.map((option, i) => {
-        return (
-          <button
-            key={i}
-            onClick={() => onClick(i)}
-            className={`${styles.btn} ${
-              selectedIndex === i ? styles.selected : ''
-            }`}
-          >
-            {option}
-          </button>
-        )
-      })}
-    </div>
-  )
-}
+/* import BaseSelect from './BaseSelect' */
 
-BaseSelect.propTypes = {
-  options: PropTypes.any.isRequired,
-  selectedIndex: PropTypes.number.isRequired,
-  onClick: PropTypes.func.isRequired
+class BaseSelect extends Component {
+  static propTypes = {
+    options: PropTypes.any.isRequired,
+    className: PropTypes.string.isRequired,
+    selectedClassName: PropTypes.string,
+    selectedIndex: PropTypes.number,
+    onSelect: PropTypes.func
+  }
+
+  static defaultProps = {
+    selectedIndex: 0,
+    onSelect: () => {},
+    selectedClassName: ''
+  }
+
+  state = {
+    selectedIndex: this.props.selectedIndex
+  }
+
+  onClick = selectedIndex => {
+    const { onSelect } = this.props
+    this.setState({ selectedIndex })
+
+    onSelect(selectedIndex)
+  }
+
+  render() {
+    const { className, selectedClassName, options } = this.props
+    const { selectedIndex } = this.state
+    return options.map((option, i) => {
+      return (
+        <button
+          key={i}
+          className={`${className} ${i === selectedIndex && selectedClassName}`}
+          onClick={() => this.onClick(i)}
+        >
+          {option}
+        </button>
+      )
+    })
+  }
 }
 
 export default BaseSelect
