+++ src/Select/index.js
@@ -1,1 +1,1 @@
-export { default } from './Switch'
+export { default } from './Select'
