+++ stories/SwitchStory.js
@@ -1,20 +1,23 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
-import Select from '../src/Select/Select'
+import HorizontalSelect from '../src/Select'
 
 import ColorModeComparison from './ColorModeComparison'
 
-storiesOf('Select', module).add('Simple', () => (
+storiesOf('HorizontalSelect', module).add('Simple', () => (
   <div>
     <ColorModeComparison>
-      <Select options={['First', 'Second']} onSelect={action('Selected')} />
-      <Select
+      <HorizontalSelect
+        options={['First', 'Second']}
+        onSelect={action('Selected')}
+      />
+      <HorizontalSelect
         options={['First', 'Second']}
         selectedIndex={1}
         onSelect={action('Selected')}
       />
-      <Select
+      <HorizontalSelect
         options={['First', 'Second', 'Third']}
         onSelect={action('Selected')}
       />
